requirements.txt
streamlit>=1.32.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
tiktoken>=0.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
numpy>=1.26.0
openai>=1.30.0
ragas>=0.2.0
datasets>=2.20.0
sentence-transformers>=3.0.0
